+++ backend/canisters/group/impl/src/updates/register_proposal_vote.rs
@@ -30,7 +30,7 @@ async fn register_proposal_vote_impl(args: Args) -> Response {
         proposal_id,
         adopt: args.adopt,
     };
-    match user_canister_c2c_client::c2c_vote_on_proposal(user_id.into(), &c2c_args).await {
+    match user_canister_c2c_client::c2c_vote_on_proposal(user_id.canister_id(), &c2c_args).await {
         Ok(response) => match response {
             user_canister::c2c_vote_on_proposal::Response::Success => mutate_state(|state| commit(user_id, args, state)).into(),
             user_canister::c2c_vote_on_proposal::Response::Error(error) => Response::Error(error),