+++ backend/integration_tests/src/suspend_user_tests.rs
@@ -48,7 +48,7 @@ fn suspend_user() {
     let direct_message_response1 = client::user::send_message_v2(
         env,
         user1.principal,
-        user1.user_id.into(),
+        user1.user_id.canister_id(),
         &user_canister::send_message_v2::Args {
             recipient: user2.user_id,
             thread_root_message_index: None,
@@ -134,7 +134,7 @@ fn suspend_user() {
     let direct_message_response2 = client::user::send_message_v2(
         env,
         user1.principal,
-        user1.user_id.into(),
+        user1.user_id.canister_id(),
         &user_canister::send_message_v2::Args {
             recipient: user2.user_id,
             thread_root_message_index: None,